fix DMB C implementation to build multi-arch
The previous C implementation of DMB only built correctly for
arm. Now this builds for any arch by detecting whether arm or aarch64
is the target arch, and setting the C -march accordingly, or using
a nop function for any other arch (which can be useful when a
crate depends on xilinx-dma for some functionality, but has some
tests that could run under x86_64 because they don't use xilinx-dma).