axi_dma: add size_d2h() method to get packet size
According to the AXI DMA documentation, the S2MM_LENGTH register
(which corresponds directly to the bytes parameter of the start_d2h()
method) is used to indicate the length in bytes of the buffer used
to hold the data to be transferred. However, the data is an AXI-Stream
packet, so its size comes determined by the TLAST line. Therefore, the
actual size of the transferred data may be smaller than the buffer size.
After completion of the transfer, the S2MM_LENGTH register is updated
with the actual number of bytes that have been transferred.

Thus, the user can call start_d2h() with the size of the buffer, then
call wait_d2h(), and afterwards call size_d2h() to obtain the length
of the data that has been transferred.

Note that the behaviour of the AXI DMA is undefined if the buffer is not
large enough to hold the AXI-Stream packet, so the user should ensure
that the buffer is large enough to hold any possible AXI-Stream packet
that the AXI DMA might receive on its S_AXIS_S2MM port.