use inline assembly for dmb
Inline assembly for arm and aarch64 has been in Rust stable for
quite some time already. This removes the C function used to perform
"dmb sy" and replaces it by inline assembly. The advantages are
that the code is simpler, does not need a C compiler, and the
"dmb sy" instruction gets inlined where it needs to be instead
of staying behind a function call.